examples/longvit/requirements.txt
torch==2.0.0
timm==0.6.13
Pillow==10.0.0
blobfile==2.0.2
mypy==1.4.1
numpy==1.22.4
pytest==7.2.2
requests==2.31.0
einops==0.6.1
tensorboardX==1.8
scipy==1.6.3
ftfy==6.1.1
opencv-python==4.8.0.74
pyarrow==9.0.0
transformers==4.8.1
deepspeed==0.4.0